requirements.txt
djangorestframework
django-filter
django-restql
django-cors-headers
django-environ
psycopg2-binary
uvicorn